pyproject.toml
[tool.poetry]
name = "weatherdata"
version = "0.1.0"
description = ""
authors = ["Alexey Lukyanov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
urllib3 = "1.26.19"
poetry = "1.4.0"
redis = "5.0.7"
numpy = "2.0.0"
plotly = "5.22.0"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"
pytz = "^2024.1"
dash = "^2.17.1"
dash-bootstrap-components = "^1.6.0"
pandas = "^2.2.2"
gunicorn = "^23.0.0"
pyarrow = "^17.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
